feat(refactor): Extract services from ProxyBlock.php
- Create helper services that respect the Single Responsibility Principle.
- Use the new services in ProxyBlock.php.
- Reduce the cyclomatic complexity of ProxyBlock.php.
